Feature/workspace settings

+++ frontend/src/Main/components/TimelineForSent/TimelineForSent.tsx
@@ -45,6 +45,7 @@ function TimelineForSent(props: timelineForSentProps) {
 
         axios(options)
             .then(() => {
+                alert("コントリビューションを削除しました。");
                 props.setContributions(() => {
                     return props.contributions.filter(
                         (contribution) => {
@@ -122,7 +123,12 @@ function TimelineForSent(props: timelineForSentProps) {
     };
 
     const onClickDeleteButton = (contributionId: string) => {
-        deleteContribution(contributionId);
+        const result = window.confirm("確認：コントリビューションを削除しますか？");
+        if (result) {
+            deleteContribution(contributionId);
+        } else {
+            alert("コントリビューションの削除をキャンセルしました。");
+        }
     };
 
     const onClickSendReactionButton = (contributionId: string) => {
@@ -142,26 +148,27 @@ function TimelineForSent(props: timelineForSentProps) {
     }, []);
 
     return (
-        <div className="TimelineForSent-container">
+        <div className="Timeline-container">
             <ul className="contribution_ul">
                 {props.contributions.map((contribution) => {
                     return (
                     <li className="contribution_li" key={contribution.contribution_id}>
                         <div className="contribution">
-                            {convertIdToName(contribution.sender_id)},
-                            {convertIdToName(contribution.receiver_id)},
-                            {contribution.message},
-                            {contribution.points},
-                            {contribution.reaction},
-                            {contribution.created_at},
-                            {contribution.update_at}
+                            {convertIdToName(contribution.sender_id)} さんから
+                            {convertIdToName(contribution.receiver_id)} さんへ
+                        </div>
+                        <div className="contribution">
+                            メッセージ:{contribution.message}
+                        </div>
+                        <div className="contribution">
+                            {contribution.points}ポイント
+                        </div>
+                        <div className="contribution">
+                            投稿:{contribution.created_at}
+                            更新:{contribution.update_at}
                         </div>
                         <button 
-                            className="reaction-button"
-                            onClick={() => onClickSendReactionButton(contribution.contribution_id)}
-                        >Good!!</button>
-                        <button 
-                            className="reaction-button"
+                            className="timeline-edit-button"
                             onClick={() => {props.setTargetContributionContent({
                                 contribution_id: contribution.contribution_id,
                                 points: contribution.points,
@@ -169,9 +176,13 @@ function TimelineForSent(props: timelineForSentProps) {
                             })}}
                         >編集</button>
                         <button 
-                            className="reaction-button"
+                            className="timeline-delete-button"
                             onClick={() => onClickDeleteButton(contribution.contribution_id)}
                         >削除</button>
+                         <button 
+                            className="reaction-button-sent"
+                            onClick={() => onClickSendReactionButton(contribution.contribution_id)}
+                        >いいね！{contribution.reaction}</button>
                     </li>
                     );
                 })}
